Break up real world logic
