+++ lib/createWebsite/makeRealWorldTable.js
@@ -1,72 +1,12 @@
-import { writeFileSync } from 'node:fs';
-import { join } from 'node:path';
-
 import { minifiers } from '../loaders/loadAllMinifiers.js';
-import { loadAllRealWorldTests } from '../loaders/loadAllRealWorldTests.js';
+import { minifyAllRealWorldTests } from '../runners/realWorld.js';
 import { ERROR } from '../constants.js';
 import {
-  durationNsToMs,
   formatMs,
   html
 } from '../helpers.js';
-import { minify } from '../minify.js';
 
 const __dirname = import.meta.dirname;
-
-const minifyAllRealWorldTests = async function () {
-  const startAll = process.hrtime.bigint();
-  const libraries = loadAllRealWorldTests();
-
-  for (const minifier of minifiers) {
-    for (const library of libraries) {
-      library.results = library.results || {};
-      library.duration = library.duration || {};
-      let result = '';
-      let error = false;
-      const start = process.hrtime.bigint();
-      try {
-        result = await minify(minifier, library.source);
-      } catch {
-        error = true;
-      }
-      const end = process.hrtime.bigint();
-      library.duration[minifier] = durationNsToMs(start, end);
-      if (error) {
-        library.results[minifier] = ERROR;
-      } else {
-        library.results[minifier] = result.length;
-      }
-    }
-  }
-  const jsonLibraries = [];
-  for (const library of libraries) {
-    jsonLibraries.push({
-      name: library.name,
-      version: library.version,
-      size: library.source.length,
-      results: {
-        ...library.results
-      }
-    });
-  }
-
-  const oututFile = join(__dirname, '..', '..', 'data', 'real-world-results.json');
-  const output = JSON.stringify(jsonLibraries, null, 2);
-  writeFileSync(oututFile, output + '\n');
-
-  const endAll = process.hrtime.bigint();
-  const duration = durationNsToMs(startAll, endAll);
-
-  console.log([
-    libraries.length,
-    'real world libraries minified by',
-    minifiers.length,
-    'minifiers in',
-    formatMs(duration) + '.'
-  ].join(' '));
-
-  return libraries;
-};
 
 const makeSizeTable = function (results) {
   const makeHeaders = function () {
@@ -351,7 +291,7 @@ const makeSizeTable = function (results) {
           ${allTotals}
         </tr>
         <tr>
-          <th>Total Percents</th>
+          <th>Percent Reduction</th>
           ${allPercents}
         </tr>
         <tr>
