+++ lib/reporters/README.md
@@ -1,0 +1,3 @@
+This folder is for logic related to generating reports or logging out summaries after tests have ran.
+
+It is not for logic related to running minifiers, or creating the website.
