+++ lib/runners/README.md
@@ -1,0 +1,3 @@
+Logic in this folder is focused on running minifiers against various test strings of CSS.
+
+It is not for generating the website, logging out information to the console, or creating reports or summaries.
